+++ apps/namadillo/src/App/Transactions/LocalStorageTransactionCard.tsx
@@ -1,11 +1,20 @@
 import { CopyToClipboardControl, Tooltip } from "@namada/components";
 import { shortenAddress } from "@namada/utils";
+import { FiatCurrency } from "App/Common/FiatCurrency";
 import { TokenCurrency } from "App/Common/TokenCurrency";
 import { AssetImage } from "App/Transfer/AssetImage";
 import { isShieldedAddress, isTransparentAddress } from "App/Transfer/common";
+import { namadaRegistryChainAssetsMapAtom } from "atoms/integrations";
+
+import { tokenPricesFamily } from "atoms/prices/atoms";
+import BigNumber from "bignumber.js";
 import clsx from "clsx";
+import { useAtomValue } from "jotai";
 import { FaLock } from "react-icons/fa";
-import { IoCheckmarkCircleOutline } from "react-icons/io5";
+import {
+  IoCheckmarkCircleOutline,
+  IoCloseCircleOutline,
+} from "react-icons/io5";
 import { twMerge } from "tailwind-merge";
 import { TransferTransactionData } from "types";
 import keplrSvg from "../../integrations/assets/keplr.svg";
@@ -24,6 +33,25 @@ const getTitle = (transferTransaction: TransferTransactionData): string => {
 export const LocalStorageTransactionCard = ({
   transaction,
 }: TransactionCardProps): JSX.Element => {
+  const namadaAssetsMap = useAtomValue(namadaRegistryChainAssetsMapAtom);
+  const namadaAsset =
+    namadaAssetsMap.data &&
+    Object.values(namadaAssetsMap.data).find(
+      (namadaAsset) => namadaAsset.symbol === transaction.asset.symbol
+    );
+
+  // Use the Namada asset address if available, otherwise try the original asset address
+  const assetAddress = namadaAsset?.address || transaction.asset.address;
+
+  const tokenPrices = useAtomValue(
+    tokenPricesFamily(assetAddress ? [assetAddress] : [])
+  );
+  const tokenPrice = assetAddress && tokenPrices.data?.[assetAddress];
+
+  // Ensure displayAmount is a BigNumber before performing calculations
+  const displayAmount = BigNumber(transaction.displayAmount);
+  const dollarAmount = tokenPrice && displayAmount.multipliedBy(tokenPrice);
+
   const renderKeplrIcon = (address: string): JSX.Element | null => {
     if (isShieldedAddress(address)) return null;
     if (isTransparentAddress(address)) return null;
@@ -31,6 +59,7 @@ export const LocalStorageTransactionCard = ({
   };
   const sender = transaction.sourceAddress;
   const receiver = transaction.destinationAddress;
+  const transactionFailed = transaction.status === "error";
   return (
     <article
       className={clsx(
@@ -40,13 +69,32 @@ export const LocalStorageTransactionCard = ({
       )}
     >
       <div className="flex items-center gap-3">
-        <i className={twMerge("text-2xl, text-success")}>
-          <IoCheckmarkCircleOutline className="ml-1 mt-0.5 w-10 h-10" />
+        <i
+          className={twMerge(
+            clsx("text-2xl", {
+              "text-success": !transactionFailed,
+              "text-fail": transactionFailed,
+            })
+          )}
+        >
+          {!transactionFailed && (
+            <IoCheckmarkCircleOutline className="ml-1 mt-0.5 w-10 h-10" />
+          )}
+          {transactionFailed && (
+            <IoCloseCircleOutline className="ml-1 mt-0.5 w-10 h-10" />
+          )}
         </i>
 
         <div className="flex flex-col">
-          <h3 className="text-success flex relative group/tooltip">
-            {getTitle(transaction)}{" "}
+          <h3
+            className={twMerge(
+              clsx("flex relative group/tooltip", {
+                "text-success": !transactionFailed,
+                "text-fail": transactionFailed,
+              })
+            )}
+          >
+            {transactionFailed && "Failed"} {getTitle(transaction)}{" "}
             <CopyToClipboardControl
               className="ml-1.5 text-neutral-400"
               value={transaction?.hash ?? ""}
@@ -71,25 +119,22 @@ export const LocalStorageTransactionCard = ({
         <div className="aspect-square w-10 h-10">
           <AssetImage asset={transaction.asset} />
         </div>
-        <TokenCurrency
-          className="text-white mt-1 ml-2"
-          amount={transaction.displayAmount}
-          symbol={transaction.asset.symbol}
-        />
+        <div className="ml-2 flex flex-col">
+          <TokenCurrency
+            className="text-white"
+            amount={displayAmount}
+            symbol={transaction.asset.symbol}
+          />
+          {dollarAmount && (
+            <FiatCurrency
+              className="text-neutral-400 text-sm"
+              amount={dollarAmount}
+            />
+          )}
+        </div>
       </div>
       <div className="flex flex-col">
-        <h4
-          className={
-            (
-              isShieldedAddress(sender ?? "") ||
-              transaction.type === "ShieldedToIbc"
-            ) ?
-              "text-yellow"
-            : ""
-          }
-        >
-          From
-        </h4>
+        <h4 className="text-neutral-400">From</h4>
         <h4
           className={
             (
@@ -116,9 +161,7 @@ export const LocalStorageTransactionCard = ({
       </div>
 
       <div className="flex flex-col relative">
-        <h4 className={isShieldedAddress(receiver ?? "") ? "text-yellow" : ""}>
-          To
-        </h4>
+        <h4 className="text-neutral-400">To</h4>
         <div className="flex items-center justify-between">
           <h4
             className={isShieldedAddress(receiver ?? "") ? "text-yellow" : ""}