+++ apps/namadillo/src/atoms/chain/services.ts
@@ -1,8 +1,8 @@
 import {
   DefaultApi,
   IbcToken,
+  ModelParameters,
   NativeToken,
-  Parameters,
 } from "@namada/indexer-client";
 import { getDenomFromIbcTrace } from "atoms/integrations";
 import BigNumber from "bignumber.js";
@@ -18,7 +18,7 @@ export const fetchRpcUrlFromIndexer = async (
 };
 
 // TODO: We need the response type of this call updated in the indexer client
-type TempChainParams = Parameters & {
+type TempChainParams = ModelParameters & {
   checksums: {
     current: Record<string, string>;
     fallback: Record<string, string>;
@@ -50,9 +50,9 @@ export const fetchMaspRewards = async (
   const existingRewards: MaspAssetRewards[] = rewards
     .filter((r) => r.maxRewardRate > 0)
     .map((r) => {
-      const denom = getDenomFromIbcTrace(r.name);
       const asset =
-        assets.find((asset) => asset.base === denom) ?? unknownAsset(denom);
+        assets.find((asset) => asset.address === r.address) ??
+        unknownAsset(getDenomFromIbcTrace(r.name));
       return {
         asset,
         address: r.address,